Book CRUD

+++ src/main/java/com/librarymanagement/controller/BookController.java
@@ -1,8 +1,11 @@
 package com.librarymanagement.controller;
 
-import com.librarymanagement.entity.Book;
+import com.librarymanagement.dto.BookDTO;
 import com.librarymanagement.service.BookService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,23 +21,39 @@ public class BookController {
     }
 
     @GetMapping
-    public List<Book> getAllBooks() {
-        return service.getAllBooks();
+    public ResponseEntity<List<BookDTO>> getAllBooks() {
+        List<BookDTO> books = service.getAllBooks();
+        return new ResponseEntity<>(books, HttpStatus.OK);
     }
 
     @PostMapping
-    public Book addBook(@RequestBody Book book) {
-        return service.addBook(book);
+    public ResponseEntity<BookDTO> addBook(@Valid @RequestBody BookDTO bookDTO) {
+        BookDTO createdBook = service.addBook(bookDTO);
+        return new ResponseEntity<>(createdBook, HttpStatus.CREATED);
     }
 
     @GetMapping("/{id}")
-    public Book getBookById(@PathVariable Long id) {
-        return service.getBookById(id);
+    public ResponseEntity<BookDTO> getBookById(@PathVariable Long id) {
+        if (id <= 0) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        BookDTO bookById = service.getBookById(id);
+        return new ResponseEntity<>(bookById, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBook(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBook(@PathVariable Long id) {
+        if (id <= 0) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         service.deleteBook(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<BookDTO> updateBook(@PathVariable Long id, @RequestBody @Valid BookDTO bookDTO) {
+        BookDTO updatedBook = service.updateBook(id, bookDTO);
+        return new ResponseEntity<>(updatedBook, HttpStatus.OK);
     }
 
 }
